fix: tighten auth and state-dir helpers
