+++ packages/adapter-openclaw/src/DkgNodePlugin.ts
@@ -45,8 +45,8 @@ import type {
   OpenClawToolResult,
 } from './types.js';
 import { homedir } from 'node:os';
-import { dirname, join, resolve } from 'node:path';
-import { existsSync, realpathSync } from 'node:fs';
+import { join } from 'node:path';
+import { canonicalPathForCompare } from './state-dir-path.js';
 
 // T44 — same regex as `dkg-core/src/dkg-home.ts` ETH_ADDR_RE. Kept
 // duplicated rather than exposed from dkg-core because the adapter's
@@ -55,24 +55,6 @@ import { existsSync, realpathSync } from 'node:fs';
 const ETH_ADDR_RE_LC = /^0x[0-9a-f]{40}$/;
 function isValidEthAddressString(value: string | undefined): boolean {
   return typeof value === 'string' && ETH_ADDR_RE_LC.test(value.trim().toLowerCase());
-}
-
-function canonicalPathForCompare(path: string): string {
-  const absolute = resolve(path);
-  const missingParts: string[] = [];
-  let existing = absolute;
-  while (!existsSync(existing)) {
-    const parent = dirname(existing);
-    if (parent === existing) break;
-    missingParts.unshift(existing.slice(parent.length + 1));
-    existing = parent;
-  }
-
-  let canonicalBase = existing;
-  try { canonicalBase = realpathSync(existing); } catch { /* keep resolved fallback */ }
-  const canonical = missingParts.reduce((acc, part) => join(acc, part), canonicalBase);
-  const normalized = resolve(canonical);
-  return process.platform === 'win32' ? normalized.toLowerCase() : normalized;
 }
 
 const OPENCLAW_LOCAL_AGENT_CAPABILITIES = {
